refactor: fix api path
